Replace test_cached with new test_experimental
This deletes the old test_cached.py and renames the current
test_experimental.py (which consists of tests reorganized from
test_cached) to test_cached.py.
