+++ tests/test_cached.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 """
-Tests for embedding functions in the ``embed.cached`` submodule.
+Tests specific to the embedding functions in the ``embed.cached`` submodule.
 
 Those embedding functions are the versions that cache to disk. They are
 otherwise like the same-named functions residing directly in ``embed``.
@@ -10,14 +10,12 @@ otherwise like the same-named functions residing directly in ``embed``.
 # pylint: disable=missing-function-docstring
 # All test methods have self-documenting names.
 
+from abc import ABC, abstractmethod
 import json
 import pathlib
 import tempfile
-from typing import Any
 import unittest
 from unittest.mock import patch
-
-from parameterized import parameterized_class
 
 import embed
 from embed import cached
@@ -46,17 +44,9 @@ def _patch_non_disk_caching_embedder(name):
     )
 
 
-@parameterized_class(('name', 'func'), [
-    (cached.embed_one.__name__, staticmethod(cached.embed_one)),
-    (cached.embed_one_eu.__name__, staticmethod(cached.embed_one_eu)),
-    (cached.embed_one_req.__name__, staticmethod(cached.embed_one_req)),
-])
 @_helpers.maybe_cache_embeddings_in_memory
-class TestDiskCachedEmbedOne(unittest.TestCase):
-    """Tests of ``embed.cached.embed_one*`` functions, which cache to disk."""
-
-    name: Any
-    func: Any
+class _TestDiskCachedEmbedBase(ABC, unittest.TestCase):
+    """Tests of ``embed.cached.embed*`` functions, which cache to disk."""
 
     def setUp(self):
         """Create a temporary directory."""
@@ -68,22 +58,51 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
         """Delete the temporary directory."""
         self._temporary_directory.cleanup()
 
+    @property
+    @abstractmethod
+    def text_or_texts(self):
+        """Input to the embedding functions."""
+
+    @property
+    @abstractmethod
+    def filename(self):
+        """
+        Filename that should be generated from the input ``text_or_texts``.
+        """
+
+    @property
+    @abstractmethod
+    def fake_data(self):
+        """Fake data for testing loads from a file."""
+
+    @property
+    @abstractmethod
+    def func_group(self):
+        """
+        Single or multiple embedding function group. Compatible with ``func``.
+        """
+
+    @property
+    @abstractmethod
+    def func(self):
+        """Disk caching embedding function being tested."""
+
     # FIXME: Test that returned embeddings could plausibly be correct.
 
     def test_calls_same_name_non_caching_version_if_not_cached(self):
-        with _patch_non_disk_caching_embedder(self.name) as mock:
-            self.func('hola', data_dir=self._dir_path)
+        with _patch_non_disk_caching_embedder(self._name) as mock:
+            self.func(self.text_or_texts, data_dir=self._dir_path)
 
-        mock.assert_called_once_with('hola')
+        mock.assert_called_once_with(self.text_or_texts)
 
     def test_saves_file_if_not_cached(self):
         expected_message = 'INFO:embed.cached:{name}: saved: {path}'.format(
-            name=self.name,
+            name=self._name,
             path=self._path,
         )
 
         with self.assertLogs(logger=cached.__name__) as log_context:
-            self.func('hola', data_dir=self._dir_path)
+            self.func(self.text_or_texts, data_dir=self._dir_path)
 
         self.assertEqual(log_context.output, [expected_message])
 
@@ -91,22 +110,20 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
         self._write_fake_data_file()
 
         expected_message = 'INFO:embed.cached:{name}: loaded: {path}'.format(
-            name=self.name,
+            name=self._name,
             path=self._path,
         )
 
         with self.assertLogs(logger=cached.__name__) as log_context:
-            self.func('hola', data_dir=self._dir_path)
+            self.func(self.text_or_texts, data_dir=self._dir_path)
 
         self.assertEqual(log_context.output, [expected_message])
 
     def test_saves_file_that_any_implementation_can_load(self):
-        self.func('hola', data_dir=self._dir_path)
+        self.func(self.text_or_texts, data_dir=self._dir_path)
         message_format = 'INFO:embed.cached:{name}: loaded: {path}'
 
-        for load_func in (cached.embed_one,
-                          cached.embed_one_eu,
-                          cached.embed_one_req):
+        for load_func in self.func_group:
             with self.subTest(load_func=load_func):
                 expected_message = message_format.format(
                     name=load_func.__name__,
@@ -114,7 +131,7 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
                 )
 
                 with self.assertLogs(logger=cached.__name__) as log_context:
-                    load_func('hola', data_dir=self._dir_path)
+                    load_func(self.text_or_texts, data_dir=self._dir_path)
 
                 self.assertEqual(log_context.output, [expected_message])
 
@@ -124,7 +141,7 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
         expected_open_event = _audit.OpenEvent(str(self._path), 'r')
 
         with _audit.listening_for_open() as open_events:
-            self.func('hola', data_dir=self._dir_path)
+            self.func(self.text_or_texts, data_dir=self._dir_path)
 
         self.assertIn(expected_open_event, open_events)
 
@@ -134,23 +151,23 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
         expected_open_event = _audit.OpenEvent(str(self._path), 'x')
 
         with _audit.listening_for_open() as open_events:
-            self.func('hola', data_dir=self._dir_path)
+            self.func(self.text_or_texts, data_dir=self._dir_path)
 
         self.assertIn(expected_open_event, open_events)
 
     def test_saved_embedding_exists(self):
-        self.func('hola', data_dir=self._dir_path)
+        self.func(self.text_or_texts, data_dir=self._dir_path)
         self.assertTrue(self._path.is_file())
 
     def test_uses_default_data_dir_if_not_passed(self):
         expected_message = 'INFO:embed.cached:{name}: saved: {path}'.format(
-            name=self.name,
+            name=self._name,
             path=self._path,
         )
 
         with patch(f'{cached.__name__}.DEFAULT_DATA_DIR', self._dir_path):
             with self.assertLogs(logger=cached.__name__) as log_context:
-                self.func('hola')
+                self.func(self.text_or_texts)
 
         self.assertEqual(
             log_context.output, [expected_message],
@@ -158,143 +175,117 @@ class TestDiskCachedEmbedOne(unittest.TestCase):
         )
 
     @property
+    def _name(self):
+        """Name of the disk caching embedding function being tested."""
+        return self.func.__name__
+
+    @property
     def _path(self):
         """Path of temporary test file."""
-        return self._dir_path / _HOLA_FILENAME
+        return self._dir_path / self.filename
 
     def _write_fake_data_file(self):
         """Create a file containing a fake embedding."""
-        fake_data = [1.0] + [0.0] * (embed.DIMENSION - 1)  # Normalized vector.
         with open(file=self._path, mode='w', encoding='utf-8') as file:
-            json.dump(obj=fake_data, fp=file)
+            json.dump(obj=self.fake_data, fp=file)
 
 
-@parameterized_class(('name', 'func'), [
-    (cached.embed_many.__name__, staticmethod(cached.embed_many)),
-    (cached.embed_many_eu.__name__, staticmethod(cached.embed_many_eu)),
-    (cached.embed_many_req.__name__, staticmethod(cached.embed_many_req)),
-])
-@_helpers.maybe_cache_embeddings_in_memory
-class TestDiskCachedEmbedMany(unittest.TestCase):
-    """Tests of ``embed.cached.embed_many*`` functions, which cache to disk."""
-
-    name: Any
-    func: Any
-
-    def setUp(self):
-        """Create a temporary directory."""
-        # pylint: disable=consider-using-with  # tearDown cleans this up.
-        self._temporary_directory = tempfile.TemporaryDirectory()
-        self._dir_path = pathlib.Path(self._temporary_directory.name)
-
-    def tearDown(self):
-        """Delete the temporary directory."""
-        self._temporary_directory.cleanup()
-
-    # FIXME: Test that returned embeddings could plausibly be correct.
-
-    def test_calls_same_name_non_caching_version_if_not_cached(self):
-        with _patch_non_disk_caching_embedder(self.name) as mock:
-            self.func(['hola', 'hello'], data_dir=self._dir_path)
-
-        mock.assert_called_once_with(['hola', 'hello'])
-
-    def test_saves_file_if_not_cached(self):
-        expected_message = 'INFO:embed.cached:{name}: saved: {path}'.format(
-            name=self.name,
-            path=self._path,
-        )
-
-        with self.assertLogs(logger=cached.__name__) as log_context:
-            self.func(['hola', 'hello'], data_dir=self._dir_path)
-
-        self.assertEqual(log_context.output, [expected_message])
-
-    def test_loads_file_if_cached(self):
-        self._write_fake_data_file()
-
-        expected_message = 'INFO:embed.cached:{name}: loaded: {path}'.format(
-            name=self.name,
-            path=self._path,
-        )
-
-        with self.assertLogs(logger=cached.__name__) as log_context:
-            self.func(['hola', 'hello'], data_dir=self._dir_path)
-
-        self.assertEqual(log_context.output, [expected_message])
-
-    def test_saves_file_that_any_implementation_can_load(self):
-        self.func(['hola', 'hello'], data_dir=self._dir_path)
-        message_format = 'INFO:embed.cached:{name}: loaded: {path}'
-
-        for load_func in (cached.embed_many,
-                          cached.embed_many_eu,
-                          cached.embed_many_req):
-            with self.subTest(load_func=load_func):
-                expected_message = message_format.format(
-                    name=load_func.__name__,
-                    path=self._path,
-                )
-
-                with self.assertLogs(logger=cached.__name__) as log_context:
-                    load_func(['hola', 'hello'], data_dir=self._dir_path)
-
-                self.assertEqual(log_context.output, [expected_message])
-
-    @_audit.skip_if_unavailable
-    def test_load_confirmed_by_audit_event(self):
-        self._write_fake_data_file()
-        expected_open_event = _audit.OpenEvent(str(self._path), 'r')
-
-        with _audit.listening_for_open() as open_events:
-            self.func(['hola', 'hello'], data_dir=self._dir_path)
-
-        self.assertIn(expected_open_event, open_events)
-
-    @_audit.skip_if_unavailable
-    def test_save_confirmed_by_audit_event(self):
-        # TODO: Decide whether to keep allowing just 'x', or if 'w' is OK too.
-        expected_open_event = _audit.OpenEvent(str(self._path), 'x')
-
-        with _audit.listening_for_open() as open_events:
-            self.func(['hola', 'hello'], data_dir=self._dir_path)
-
-        self.assertIn(expected_open_event, open_events)
-
-    def test_saved_embedding_exists(self):
-        self.func(['hola', 'hello'], data_dir=self._dir_path)
-        self.assertTrue(self._path.is_file())
-
-    def test_uses_default_data_dir_if_not_passed(self):
-        expected_message = 'INFO:embed.cached:{name}: saved: {path}'.format(
-            name=self.name,
-            path=self._path,
-        )
-
-        with patch(f'{cached.__name__}.DEFAULT_DATA_DIR', self._dir_path):
-            with self.assertLogs(logger=cached.__name__) as log_context:
-                self.func(['hola', 'hello'])
-
-        self.assertEqual(
-            log_context.output, [expected_message],
-            'DEFAULT_DATA_DIR should be used',
-        )
+class _TestDiskCachedEmbedOneBase(_TestDiskCachedEmbedBase):
+    """Abstract base for ``embed_one*`` group customizations."""
 
     @property
-    def _path(self):
-        """Path of temporary test file."""
-        return self._dir_path / _HOLA_HELLO_FILENAME
+    def text_or_texts(self):
+        return 'hola'
 
-    def _write_fake_data_file(self):
-        """Create a file containing a fake embedding."""
-        # Two normalized vectors.
-        fake_data = [
+    @property
+    def filename(self):
+        return _HOLA_FILENAME
+
+    @property
+    def fake_data(self):
+        """Normalized vector."""
+        return [1.0] + [0.0] * (embed.DIMENSION - 1)  # Normalized vector.
+
+    @property
+    def func_group(self):
+        return cached.embed_one, cached.embed_one_eu, cached.embed_one_req
+
+
+class _TestDiskCachedEmbedManyBase(_TestDiskCachedEmbedBase):
+    """Abstract base for ``embed_many*`` group customizations."""
+
+    @property
+    def text_or_texts(self):
+        return ['hola', 'hello']
+
+    @property
+    def filename(self):
+        return _HOLA_HELLO_FILENAME
+
+    @property
+    def fake_data(self):
+        """Two normalized vectors."""
+        return [
             [1.0] + [0.0] + [0.0] * (embed.DIMENSION - 2),
             [0.0] + [1.0] + [0.0] * (embed.DIMENSION - 2),
         ]
 
-        with open(file=self._path, mode='w', encoding='utf-8') as file:
-            json.dump(obj=fake_data, fp=file)
+    @property
+    def func_group(self):
+        return cached.embed_many, cached.embed_many_eu, cached.embed_many_req
+
+
+class TestDiskCachedEmbedOne(_TestDiskCachedEmbedOneBase):
+    """Tests for disk cached ``embed_one``."""
+
+    @property
+    def func(self):
+        return cached.embed_one
+
+
+class TestDiskCachedEmbedOneEu(_TestDiskCachedEmbedOneBase):
+    """Tests for disk cached ``embed_one_eu``."""
+
+    @property
+    def func(self):
+        return cached.embed_one_eu
+
+
+class TestDiskCachedEmbedOneReq(_TestDiskCachedEmbedOneBase):
+    """Tests for disk cached ``embed_one_req``."""
+
+    @property
+    def func(self):
+        return cached.embed_one_req
+
+
+class TestDiskCachedEmbedMany(_TestDiskCachedEmbedManyBase):
+    """Tests for disk cached ``embed_many``."""
+
+    @property
+    def func(self):
+        return cached.embed_many
+
+
+class TestDiskCachedEmbedManyEu(_TestDiskCachedEmbedManyBase):
+    """Tests for disk cached ``embed_many_eu``."""
+
+    @property
+    def func(self):
+        return cached.embed_many_eu
+
+
+class TestDiskCachedEmbedManyReq(_TestDiskCachedEmbedManyBase):
+    """Tests for disk cached ``embed_many_req``."""
+
+    @property
+    def func(self):
+        return cached.embed_many_req
+
+
+del _TestDiskCachedEmbedBase
+del _TestDiskCachedEmbedOneBase
+del _TestDiskCachedEmbedManyBase
 
 
 if __name__ == '__main__':
